% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzupdater.R
\name{get_active_tz_db}
\alias{get_active_tz_db}
\title{get_active_tz_db}
\usage{
get_active_tz_db()
}
\value{
Active tz database for the current session.
}
\description{
Get the active tz database version
}
\examples{
# Get the active tz database in your session:
get_active_tz_db()
}
