% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzupdater.R
\name{get_last_published_tz}
\alias{get_last_published_tz}
\title{get_last_published_tz}
\usage{
get_last_published_tz()
}
\value{
Latest IANA tz db version name or Unknown if not found.
}
\description{
Get the name of the latest version available at IANA website.
}
\examples{
# Will return for instance "2019c"
get_last_published_tz()
}
